Change engine model

+++ Car-Dashboard-WPF/ModelInterfaceCommunication.cs
@@ -4,7 +4,7 @@ namespace Car_Dashboard_WPF
 {
     class ModelInterfaceCommunication
     {
-        public int currentSpeed { get; set; }
+        public double currentSpeed { get; set; }
 
         public void UpdateCurrentValues(EngineModel engine)
         {
